requirements.txt
pymem>=1.13.0

PyQt6>=6.5.0

pyyaml>=6.0
loguru>=0.7.0

keyboard>=0.13.5
mouse>=0.7.1

